add Timeline class + rspec tests
Adds a Timeline class that allows for work dates to be recorded for a
set of projects in an iterative fashion.  Timeline applies date ordering
to the set of dates worked, provides methods for identifying whether
individual dates are considered 'travel' or 'full' days, along with some
other minor quality of life features.

Includes spec tests.